requirements/dev.txt
coverage==5.5
pytest~=6.1
pytest-cov~=2.12
pytest-django~=4.5
selenium~=3.141
tox~=3.24
